Dry run of package installation

Dry run of package installation.

conan_dryrun(
  packages,
  policy = "upgrade",
  repos = NULL,
  cran = NULL,
  lib = NULL,
  error = TRUE
)

Arguments

packages
A character vector of packages to install. These can be names of CRAN packages or GitHub references etc; see pkgdepends::new_pkg_installation_proposal() for more details

policy
Should be either "lazy" or "upgrade", with a default of "upgrade"

repos
A character vector of repositories to use when installing. A suitable CRAN repo will be added if not detected (using the cran argument if provided)

cran
Fallback CRAN repo to use. If not given we will use https://cloud.r-project.org

lib
The path to install into. We will create a self-contained library at this address.

error
Logical, indicating if error should be thrown on failure. If FALSE then we return the proposal object instead.

Value

Invisibly, the resolution

Examples

conan::conan_dryrun("cpp11")

#> 

#> ✔ Updated metadata database: 4.45 MB in 12 files.

#> 

#> ℹ Updating metadata database

#> ✔ Updating metadata database ... done

#> 

#> <pkg_installation_proposal>
#> + refs:
#>   - cpp11
#> + solution policy: upgrade
#> + has solution
#> (use `$download()` to download packages)
#> (use `$show_solution()` to see the packages to install
#> (use `$get_solution()` to see the full solution results)
#> (use `$draw()` to draw the dependency tree)
#> (use `$create_lockfile()` to write a lock file)