Watch a conan installation

Watch a conan installation from another process. This can be used to stream the status and progress back to the current R process.

conan_watch(
  get_status,
  get_log,
  show_progress = TRUE,
  show_log = TRUE,
  poll = 1,
  error = TRUE
)

Arguments

get_status
A callback to get the status of the installation. The callback will be called with no arguments and must return COMPLETE on successful completion, ERROR on failure and any other value to continue.

get_log
A callback to read logs of the installation (something like function() readLines(filename, warn = FALSE) may be sufficient)

show_progress
Logical, indicating if we should show a progress bar. This requires the progress package.

show_log
Logical, indicating if the installation log should be printed

poll
Time, in seconds, used to throttle calls to the status function. The default is 1 second

error
Logical, indicating if we should throw an error if installation fails.

Value

The final status (COMPLETE or ERROR)